Moves scroll listener to data cell render.
Adds counter state to trigger rerender on scroll or intersectObserver cahnges.
